+++ MenuTracker/src/main/java/com/project/MenuTracker/service/testService.java
@@ -1,0 +1,4 @@
+package com.project.MenuTracker.service;
+
+public class testService {
+}
