+++ MenuTracker/src/main/resources/swagger/swagger.yaml
